Split HEOM bath descriptions out into their own module.
